% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airlines}
\alias{airlines}
\title{Airline names}
\format{
A data frame with columns:
\describe{
 \item{carrier}{The variable is character. Two letter abbreviation}
 \item{name}{The variable is character. Full name}
}
}
\source{
\url{https://github.com/hadley/nycflights13}
}
\usage{
airlines
}
\description{
Look up airline names from their carrier codes.
}
\examples{
\dontrun{
airlines
}
}
\keyword{datasets}
