% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airports}
\alias{airports}
\title{Airport metadata}
\format{
A data frame with columns:
\describe{
 \item{faa}{The variable is character. FAA airport code}
 \item{name}{The variable is character. Usual name of the airport}
 \item{lat}{The variable is numeric. Lattitude of the airport location}
 \item{lon}{The variable is numeric. Longitude of the airport location}
 \item{alt}{The variable is numeric. Altitude in feet}
 \item{tz}{The variable is numeric. Timezone offset from GMT}
 \item{dst}{The variable is character. Daylight savings time zone. A = Standard US DST: starts on the second Sunday of March, ends on the first Sunday of November. U = unknown. N = no dst.}
 \item{tzone}{The variable is character. IANA time zone, as determined by GeoNames webservice.}
}
}
\source{
\url{https://github.com/hadley/nycflights13}
}
\usage{
airports
}
\description{
Useful metadata about airports.
}
\examples{
\dontrun{
airports
}
}
\keyword{datasets}
