% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{batting}
\alias{batting}
\title{Batting table}
\format{
A data frame with columns:
\describe{
 \item{playerID}{The variable is character. Player ID code}
 \item{yearID}{The variable is integer. Year}
 \item{stint}{The variable is integer. player's stint (order of appearances within a season)}
 \item{teamID}{The variable is factor. Team}
 \item{lgID}{The variable is factor. League; with values AA AL FL NL PL UA}
 \item{G}{The variable is integer. Games: number of games in which a player played}
 \item{AB}{The variable is integer. At Bats}
 \item{R}{The variable is integer. Runs}
 \item{H}{The variable is integer. Hits: times reached base because of a batted, fair ball without error by the defense}
 \item{X2B}{The variable is integer. Doubles: hits on which the batter reached second base safely}
 \item{X3B}{The variable is integer. Triples: hits on which the batter reached third base safely}
 \item{HR}{The variable is integer. Homeruns}
 \item{RBI}{The variable is integer. Runs Batted In}
 \item{SB}{The variable is integer. Stolen Bases}
 \item{CS}{The variable is integer. Caught Stealing}
 \item{BB}{The variable is integer. Base on Balls}
 \item{SO}{The variable is integer. Strikeouts}
 \item{IBB}{The variable is integer. Intentional walks}
 \item{HBP}{The variable is integer. Hit by pitch}
 \item{SH}{The variable is integer. Sacrifice hits}
 \item{SF}{The variable is integer. Sacrifice flies}
 \item{GIDP}{The variable is integer. Grounded into double plays}
}
}
\source{
\url{https://github.com/cdalzell/Lahman}
}
\usage{
batting
}
\description{
Batting table - batting statistics
}
\examples{
\dontrun{
batting
}
}
\keyword{datasets}
