% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{faithful}
\alias{faithful}
\title{Old Faithful Geyser Data}
\format{
A data frame with columns:
\describe{
 \item{eruptions}{The variable is numeric. Eruption time in minutes}
 \item{waiting}{The variable is numeric. Waiting time to next eruption}
}
}
\source{
\url{https://doi.org/10.2307/2347385}
}
\usage{
faithful
}
\description{
Waiting time between eruptions and the duration of the eruption for the Old Faithful geyser in Yellowstone National Park, Wyoming, USA.
}
\examples{
\dontrun{
faithful
}
}
\keyword{datasets}
