% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fruit}
\alias{fruit}
\title{Sample strings of fruits for practicing string manipulations}
\format{
A data frame with columns:
\describe{
 \item{name}{The variable is character. Names of fruit}
}
}
\source{
\url{https://github.com/dariusk/corpora}
}
\usage{
fruit
}
\description{
fruit comes from the rcorpora package written by Gabor Csardi; the data was collected by Darius Kazemi and made available at https://github.com/dariusk/corpora.
}
\examples{
\dontrun{
fruit
}
}
\keyword{datasets}
