% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mpg}
\alias{mpg}
\title{Fuel economy data from 1999 to 2008 for 38 popular models of cars}
\format{
A data frame with columns:
\describe{
 \item{manufacturer}{The variable is character. manufacturer name}
 \item{model}{The variable is character. model name}
 \item{displ}{The variable is numeric. engine displacement, in litres}
 \item{year}{The variable is integer. year of manufacture}
 \item{cyl}{The variable is integer. number of cylinders}
 \item{trans}{The variable is character. type of transmission}
 \item{drv}{The variable is character. the type of drive train, where f = front-wheel drive, r = rear wheel drive, 4 = 4wd}
 \item{cty}{The variable is integer. city miles per gallon}
 \item{hwy}{The variable is integer. highway miles per gallon}
 \item{fl}{The variable is character. fuel type}
 \item{class}{The variable is character. 'type' of car}
}
}
\source{
\url{https://github.com/tidyverse/ggplot2}
}
\usage{
mpg
}
\description{
This dataset contains a subset of the fuel economy data that the EPA makes available on http://fueleconomy.gov. It contains only models which had a new release every year between 1999 and 2008 - this was used as a proxy for the popularity of the car.
}
\examples{
\dontrun{
mpg
}
}
\keyword{datasets}
