% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{planes}
\alias{planes}
\title{Plane metadata}
\format{
A data frame with columns:
\describe{
 \item{tailnum}{The variable is character. Tail number}
 \item{year}{The variable is integer. Year manufactured}
 \item{type}{The variable is character. }
 \item{manufacturer}{The variable is character. Manufacturer}
 \item{model}{The variable is character. model}
 \item{engines}{The variable is integer. Number of engines}
 \item{seats}{The variable is integer. Number of seats}
 \item{speed}{The variable is integer. Average cruising speed in mph}
 \item{engine}{The variable is character. Type of engine}
}
}
\source{
\url{https://github.com/hadley/nycflights13}
}
\usage{
planes
}
\description{
Plane metadata for all plane tailnumbers found in the FAA aircraft registry. American Airways (AA) and Envoy Air (MQ) report fleet numbers rather than tail numbers so can't be matched.
}
\examples{
\dontrun{
planes
}
}
\keyword{datasets}
