% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sentences}
\alias{sentences}
\title{Sample strings of sentences for practicing string manipulations}
\format{
A data frame with columns:
\describe{
 \item{name}{The variable is character. Harvard sentences}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Harvard_sentences}
}
\usage{
sentences
}
\description{
the Harvard sentences, which were designed to test VOIP systems, but are also useful for practicing regexps.
}
\examples{
\dontrun{
sentences
}
}
\keyword{datasets}
