% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weather}
\alias{weather}
\title{Hourly weather data}
\format{
A data frame with columns:
\describe{
 \item{origin}{The variable is character. Weather station. Named origin to facilitate merging with flights data.}
 \item{year}{The variable is integer. year}
 \item{month}{The variable is integer. month}
 \item{day}{The variable is integer. day}
 \item{hour}{The variable is integer. hour}
 \item{temp}{The variable is numeric. Temperature in F}
 \item{dewp}{The variable is numeric. Dewpoint in F}
 \item{humid}{The variable is numeric. Relative humidity}
 \item{wind_dir}{The variable is numeric. Wind direction}
 \item{wind_speed}{The variable is numeric. Wind speed in mph}
 \item{wind_gust}{The variable is numeric. Wind gust speed in mph}
 \item{precip}{The variable is numeric. Precipitation in inches}
 \item{pressure}{The variable is numeric. Sea level pressure in millibars}
 \item{visib}{The variable is numeric. Visibility in miles}
 \item{time_hour}{The variable is POSIXct,POSIXt. Date and hour of the recording as a POSIXct date.}
}
}
\source{
\url{https://github.com/hadley/nycflights13}
}
\usage{
weather
}
\description{
Hourly meterological data for LGA, JFK and EWR.
}
\examples{
\dontrun{
weather
}
}
\keyword{datasets}
