% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{words}
\alias{words}
\title{Sample strings of words for practicing string manipulations}
\format{
A data frame with columns:
\describe{
 \item{name}{The variable is character. words}
}
}
\source{
\url{https://github.com/dariusk/corpora}
}
\usage{
words
}
\description{
words comes from the rcorpora package written by Gabor Csardi; the data was collected by Darius Kazemi and made available at https://github.com/dariusk/corpora.
}
\examples{
\dontrun{
words
}
}
\keyword{datasets}
